+++ rest_timer.h
@@ -3,7 +3,7 @@
 
 extern bool restTimerUsingLeds;
 
-void process_rest_timer(int lastKeypressTimestamp);
+void process_rest_timer(unsigned long lastKeypressTimestamp);
 void start_break();
 
 #endif //REST_TIMER_H
